+++ package.json
@@ -7,20 +7,40 @@
     "sapper:dev": "sapper dev --ext '.svelte .md'",
     "dev": "run-p --race sapper:dev tailwind:watch",
     "copyFolder": "./scripts/copy_folder",
-    "useCache": "yarn copyFolder --merge /opt/build/cache/hc_my_cache/static/g --to static/g",
-    "fillCache": "yarn copyFolder --save static/g --to /opt/build/cache/hc_my_cache/static/g",
-    "build": "yarn clean && yarn tailwind:prod && sapper build --legacy --ext '.svelte .md'",
-    "export": "yarn tailwind:prod && NODE_ENV=production sapper export --legacy --ext '.svelte .md' --entry entries",
-    "netlify": "yarn useCache && yarn export && yarn fillCache",
+    "useCache": "npm run copyFolder -- --merge /opt/build/cache/hc_my_cache/static/g --to static/g",
+    "fillCache": "npm run copyFolder -- --save static/g --to /opt/build/cache/hc_my_cache/static/g",
+    "build": "npm run clean && npm run tailwind:prod && sapper build --legacy --ext '.svelte .md'",
+    "export": "npm run tailwind:prod && NODE_ENV=production sapper export --legacy --ext '.svelte .md' --entry entries",
+    "netlify": "npm run useCache && npm run export && npm run fillCache",
     "start": "node __sapper__/build",
     "start:exported": "npx serve __sapper__/export",
     "cy:run": "cypress run",
     "cy:open": "cypress open",
     "tailwind": "postcss static/tailwind.css -o static/index.css",
-    "tailwind:watch": "yarn tailwind -w",
-    "tailwind:prod": "NODE_ENV=production yarn tailwind",
+    "tailwind:watch": "npm run tailwind -- -w",
+    "tailwind:prod": "NODE_ENV=production npm run tailwind",
     "validate": "svelte-check --ignore src/node_modules/@sapper",
-    "test": "run-p --race dev cy:run"
+    "test": "jest src",
+    "test:serve": "run-p --race dev cy:run"
+  },
+  "jest": {
+    "transform": {
+      "^.+\\.[jt]s$": "babel-jest",
+      "^.+\\.svelte$": [
+        "svelte-jester",
+        {
+          "preprocess": true
+        }
+      ]
+    },
+    "moduleFileExtensions": [
+      "js",
+      "ts",
+      "svelte"
+    ],
+    "setupFilesAfterEnv": [
+      "@testing-library/jest-dom/extend-expect"
+    ]
   },
   "dependencies": {
     "compression": "^1.7.1",
@@ -35,6 +55,7 @@
     "@babel/plugin-syntax-dynamic-import": "^7.0.0",
     "@babel/plugin-transform-runtime": "^7.0.0",
     "@babel/preset-env": "^7.0.0",
+    "@babel/preset-typescript": "^7.12.1",
     "@babel/runtime": "^7.0.0",
     "@fullhuman/postcss-purgecss": "^1.3.0",
     "@rollup/plugin-babel": "^5.0.0",
@@ -44,14 +65,18 @@
     "@rollup/plugin-replace": "^2.2.0",
     "@rollup/plugin-typescript": "^6.0.0",
     "@rollup/plugin-url": "^5.0.0",
+    "@testing-library/jest-dom": "^5.11.6",
+    "@testing-library/svelte": "^3.0.0",
     "@tsconfig/svelte": "^1.0.10",
     "@types/compression": "^1.7.0",
+    "@types/jest": "^26.0.15",
     "@types/lodash": "^4.14.136",
-    "@types/markdown-it": "^0.0.8",
     "@types/node": "^14.11.1",
     "@types/polka": "^0.5.1",
+    "@types/testing-library__jest-dom": "^5.9.5",
     "autoprefixer": "^9.7.4",
     "babel-eslint": "^10.0.2",
+    "babel-jest": "^26.6.3",
     "cssnano": "^4.1.10",
     "eslint": "^6.1.0",
     "eslint-config-prettier": "^6.0.0",
@@ -59,6 +84,7 @@
     "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-svelte3": "^2.7.3",
     "frontmatter": "0.0.3",
+    "jest": "^26.6.3",
     "markdown-it": "^9.1.0",
     "mdsvex": "^0.8.8",
     "npm-run-all": "^4.1.5",
@@ -72,15 +98,15 @@
     "svelte": "^3.17.3",
     "svelte-check": "^1.0.46",
     "svelte-image": "happycollision/svelte-image#5952f2df",
+    "svelte-jester": "^1.1.5",
     "svelte-preprocess": "^4.3.0",
     "tailwindcss": "^1.2.0",
-    "ts-node": "^8.3.0",
     "tslib": "^2.0.1",
     "typescript": "^4.0.3",
     "xxhash": "^0.3.0"
   },
   "volta": {
     "node": "12.16.3",
-    "yarn": "1.22.4"
+    "npm": "7.0.11"
   }
 }
