+++ typings.d.ts
@@ -1,6 +1,3 @@
 type IAnyFunction = (...args: unknown[]) => unknown
 
-declare module "@sapper/server" {
-  const middleware: IAnyFunction
-  export { middleware }
-}
+type IHash<T> = Record<string, T>
